:lipstick: use micro instead of hapi
Pretty huge commit. I switched to micro as hapi is way too bloated for a micro-service, and I must admit I wanted to test it 😸.
I also use redirect again instead of proxying badge requests from shields.io. It was
originally done to avoid camo/fastly caching the image, but there is a way to force no cache, and it alleviate server load, so 👌.

- :nut_and_bolt: (heroku) no need for --no-babel
- :nut_and_bolt: (package) drop node 4